Split off names
